requirements.txt
ultralytics
opencv-python
Pillow
piexif
torch
psutil
